Center the map to include all fetched restaurants
The map was centered around user's current location. In sparse areas,
this can be confusing as there might not be any restaurants nearby which
may lead the user to believe that no restaurants were fetched.

Instead, the map could be centered in a way that all the fetched
restaurants fit the viewport.